1. Fixed some bags (initial conditions match now) 2. Created Clion project for writing code and not parallel debugging

+++ include/utils.h
@@ -14,7 +14,7 @@
 EXTERNAL double* Div_Im DEFAULT(NULL);
 
 
-double FillMask(int localNn);
+int FillMask(int localNn);
 double EnergyCalc(double k,double kmax);
 void ResetUA();
 void Symmetrize();
